requirements.txt
alembic==1.11.3
autopep8==2.0.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
platformdirs==3.2.0
pluggy==1.2.0
pycodestyle==2.10.0
PyJWT==2.7.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
virtualenv==20.22.0
Werkzeug==2.3.6
